requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
gunicorn==23.0.0
Flask-SQLAlchemy==3.1.1
cloudinary==1.44.1
pillow==11.3.0
# Dev / testing tools
pytest==8.4.2
pytest-cov==7.0.0

# HTTP clients (if needed)
httpx==0.28.1

flask-cors==6.0.1


python-multipart==0.0.20
chromadb==1.1.0
python-dotenv==1.1.1

# DB (Postgres driver) + ORM & migrations
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43
alembic==1.16.5

# Password hashing
passlib==1.7.4
bcrypt==4.3.0

openai==1.108.0

requests==2.32.5
email-validator==2.3.0

requests-oauthlib==2.0.0


langchain==0.3.27
langchain-openai==0.3.33
langchain-community==0.3.29
langgraph==0.6.7

Werkzeug==3.1.3
starlette-session==0.4.3
itsdangerous==2.2.0

python-jose==3.5.0
cryptography==46.0.1

